pyproject.toml
[tool.poetry]
name = "flask-learn"
version = "0.1.0"
description = ""
authors = ["Nic11371 <nic11371@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
